Made a rename that annoyed my for a long time.
